backend/requirements.txt
coreapi==2.3.3
coverage==5.1
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
factory-boy==2.12.0
